+++ lib/cinder/theme.ex
@@ -159,6 +159,12 @@ defmodule Cinder.Theme do
       |> apply_theme_property_mapping()
       |> apply_theme_data_attributes()
 
+  def merge("smart"),
+    do:
+      Cinder.Themes.Smart.resolve_theme()
+      |> apply_theme_property_mapping()
+      |> apply_theme_data_attributes()
+
   def merge(nil),
     do: default() |> apply_theme_property_mapping() |> apply_theme_data_attributes()